final touches

+++ src/main/java/com/example/task_manager/service/impl/TaskServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.task_manager.service.impl;
 
 import com.example.task_manager.dto.TaskDto;
+import com.example.task_manager.exceptions.NotFoundException;
 import com.example.task_manager.models.TaskEntity;
 import com.example.task_manager.models.UserEntity;
 import com.example.task_manager.repository.TaskRepository;
@@ -17,7 +18,7 @@ import java.util.stream.Collectors;
 
 @Service
 public class TaskServiceImpl implements TaskService {
-    private final TaskRepository taskRepository;
+    private TaskRepository taskRepository;
     private UserRepository userRepository;
 
     @Autowired
@@ -28,21 +29,18 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public List<TaskDto> getTasks() {
-        Object principle = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        String username;
-        if(principle instanceof UserDetails){
-            username = ((UserDetails) principle).getUsername();
-        }else {
-            username = principle.toString();
-        }
+        String username = getUsernameFromSecurity();
         UserEntity user = userRepository.findByUsername(username).orElseThrow(()->new RuntimeException("Username not found"));
         return taskRepository.findByUserId(user.getId()).stream().map(taskEntity -> mapToDto(taskEntity)).collect(Collectors.toList());
     }
 
     @Override
     public TaskDto addTask(TaskDto taskDto) {
+        String username = getUsernameFromSecurity();
+        UserEntity user = userRepository.findByUsername(username).orElseThrow(()->new NotFoundException("Username not found"));
+
         TaskEntity newTask = mapToEntity(taskDto);
-        newTask.setUserId(2);
+        newTask.setUserId(user.getId());
         newTask.setStatus(false);
         newTask.setCreatedAt(new Date());
 
@@ -53,7 +51,15 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public TaskDto updateTask(int taskId, TaskDto taskDto) {
-        TaskEntity taskEntity = taskRepository.findById(taskId).orElseThrow(()->new RuntimeException("Task not found"));
+        String username = getUsernameFromSecurity();
+        UserEntity user = userRepository.findByUsername(username).orElseThrow(()->new NotFoundException("Username not found"));
+
+        TaskEntity taskEntity = taskRepository.findById(taskId).orElseThrow(()->new NotFoundException("Task not found"));
+
+        if(user.getId() != taskEntity.getUserId()){
+            throw new NotFoundException("The task is not for the logged in user");
+        }
+
         taskEntity.setName(taskDto.getName());
 
         taskRepository.save(taskEntity);
@@ -62,14 +68,30 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public String deleteTask(int taskId) {
-        TaskEntity taskEntity = taskRepository.findById(taskId).orElseThrow(()-> new RuntimeException("Task not found"));
+        String username = getUsernameFromSecurity();
+        UserEntity user = userRepository.findByUsername(username).orElseThrow(()->new NotFoundException("Username not found"));
+
+        TaskEntity taskEntity = taskRepository.findById(taskId).orElseThrow(()-> new NotFoundException("Task not found"));
+
+        if(user.getId() != taskEntity.getUserId()){
+            throw new NotFoundException("The task is not for the logged in user");
+        }
+
         taskRepository.delete(taskEntity);
         return "Task deleted";
     }
 
     @Override
     public TaskDto markTask(int taskId) {
-        TaskEntity taskEntity = taskRepository.findById(taskId).orElseThrow(()-> new RuntimeException("Task not found"));
+        String username = getUsernameFromSecurity();
+        UserEntity user = userRepository.findByUsername(username).orElseThrow(()->new NotFoundException("Username not found"));
+
+        TaskEntity taskEntity = taskRepository.findById(taskId).orElseThrow(()-> new NotFoundException("Task not found"));
+
+        if(user.getId() != taskEntity.getUserId()){
+            throw new NotFoundException("The task is not for the logged in user");
+        }
+
         taskEntity.setStatus(!taskEntity.isStatus());
 
         taskRepository.save(taskEntity);
@@ -92,5 +114,17 @@ public class TaskServiceImpl implements TaskService {
         taskEntity.setStatus(taskDto.isStatus());
 
         return taskEntity;
+    }
+
+    private String getUsernameFromSecurity(){
+        Object principle = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username;
+        if(principle instanceof UserDetails){
+            username = ((UserDetails) principle).getUsername();
+        }else {
+            username = principle.toString();
+        }
+
+        return username;
     }
 }
